Added a Third Font Size
We now have a 4x6 sized pixel art font which should be easier on the eyes (and just looks better in generate). I have replaced many of the larger and smaller fonts with this new medium font. I might keep the smaller font in some places, but for the most part I'm changing to this one