Stripped UI and Added Busy System
Removed a lot of uneccessary persistent references in favor of accessing them with unique names when needed. Most important cases were the condition and wave labels, and the hp bars.

Also implemented the busy system which allows for entities in the process of dying to actually die before next moves are performed. This also allowed me to remove the weird waiting check I had to do before loading new waves