Reworked Tooltips + Allowed Equip Switching
Tooltips have been taken out of the global scope (something I'm trying to do more of). The core code has been placed into a much smaller class that should be inherited from. This will allow for much more flexibility when it comes to tooltips.

It's my goal to eventually change how information is displayed throughout the game. Specifically the action items, status effects, and item displays. Other info may also have this change applied when I get around to it.

I also updated the menu to add an equipment selector that updates the player's currently equipped gear. This also updates in battle and in the preview.