Added Crit System
Added the ability for moves to crit. Also added the capacity for multiple crits to land (when the percent is greater than 100%).

Also updated the Damage Marker to be much more descriptive. It now shows whether or not a move critted (and how many crits occurred). And there's also varied text color based on the elemental resistance